Added POST support for `/bugs/(id)/notes` paths, as set forth in #15.
I had to alter '''Bugnote''' significantly to take advantage of its resource-to-mantis conversion capabilities without a bugnote being the object of the request.  I think the change will be good for the project; it decouples populating the '''Bugnote''' object from the process of answering an HTTP request, with little to no loss of clarity.

Also, a few lines in `mantis_rest_core.php` used `else if` instead of `elseif`.